fix test and build
