+++ modules/web3/types/contracts.d.ts
@@ -1,4 +1,4 @@
-import { GoerliSdk, MainnetSdk } from '@dethcrypto/eth-sdk-client';
+import { GoerliSdk, MainnetSdk, ArbitrumGoerliTestnetSdk, ArbitrumOneSdk } from '@dethcrypto/eth-sdk-client';
 import { providers, Signer } from 'ethers';
 
 export type ContractName =
@@ -26,4 +26,9 @@ export type SignerOrProvider = Signer | providers.Provider;
 export type SdkGenerators = {
   mainnet: (signerOrProvider: SignerOrProvider) => MainnetSdk;
   goerli: (signerOrProvider: SignerOrProvider) => GoerliSdk;
+};
+
+export type ArbitrumSdkGenerators = {
+  mainnet: (signerOrProvider: SignerOrProvider) => ArbitrumOneSdk;
+  goerli: (signerOrProvider: SignerOrProvider) => ArbitrumGoerliTestnetSdk;
 };
