+++ pages/api/polling/vote.ts
@@ -16,13 +16,7 @@ import { MIN_MKR_REQUIRED_FOR_GASLESS_VOTING } from 'modules/polling/polling.con
 import { postRequestToDiscord } from 'modules/app/api/postRequestToDiscord';
 import { isSupportedNetwork } from 'modules/web3/helpers/networks';
 import { ballotIncludesAlreadyVoted } from 'modules/polling/helpers/ballotIncludesAlreadyVoted';
-import { Contract } from 'ethers';
-import { DefenderRelayProvider, DefenderRelaySigner } from 'defender-relay-client/lib/ethers';
-import { SupportedNetworks } from 'modules/web3/constants/networks';
-import {
-  arbitrumSdkGenerators,
-  relayerCredentials
-} from 'modules/polling/helpers/getArbitrumPollingContract';
+import { getArbitrumPollingContract } from 'modules/polling/helpers/getArbitrumPollingContract';
 
 export const API_VOTE_ERRORS = {
   VOTER_MUST_BE_STRING: 'voter must be a string',
@@ -56,18 +50,6 @@ async function postError(error: string) {
     logger.error(err);
   }
 }
-
-//we get an error if we try to run this defender relay code on the frontend
-const getArbitrumPollingContract = (network: SupportedNetworks): Contract => {
-  const sdkNetwork = network === SupportedNetworks.GOERLIFORK ? SupportedNetworks.GOERLI : network;
-  const provider = new DefenderRelayProvider(relayerCredentials[sdkNetwork]);
-  const signer = new DefenderRelaySigner(relayerCredentials[sdkNetwork], provider, {
-    speed: 'fast'
-  });
-  const { polling } = arbitrumSdkGenerators[sdkNetwork](signer);
-
-  return polling;
-};
 
 //TODO: add swagger documentation
 export default withApiHandler(
